requirements.txt
astroid==2.2.5
attrs==19.3.0
blis==0.4.1
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
colorama==0.4.1
cymem==2.0.3
docx2txt==0.8
en-core-web-sm==2.2.5
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
fpdf==1.7.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.2.0
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
murmurhash==1.0.2
nltk==3.5
numpy==1.18.2
pandas==1.0.3
pdfminer.six==20200517
plac==1.1.3
postgres==3.0.0
preshed==3.0.2
psycopg2-binary==2.8.5
psycopg2-pool==1.1
pycryptodome==3.9.7
pylint==2.3.1
pyPdf==1.13
PyPDF2==1.26.0
pyresparser==1.0.6
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
regex==2020.4.4
requests==2.23.0
six==1.12.0
sortedcontainers==2.1.0
spacy==2.2.4
SQLAlchemy==1.3.16
srsly==1.0.2
thinc==7.4.0
tqdm==4.45.0
typed-ast==1.4.0
urllib3==1.25.9
wasabi==0.6.0
Werkzeug==1.0.1
wincertstore==0.2
zipp==3.1.0
